puncturing: Add polar code specific puncturing
With this commit, systematic polar codes are properly punctured. This
should improve performance in case of sensible configurations.

puncturing: Update YAML files and examples

Puncturing is already available. Now, we can use it from within GRC.

integration: Reuse MAC CRC

With this commit we reuse a possible MAC CRC. You'll notice it failing
if the decoder time increases significantly.

pybind: Switch from SWIG to PyBind11

This module is particularly difficult to port because the
generic_encoder and generic_decoder API resides in gnuradio.fec and this
requires more includes.